@@ -4,7 +4,7 @@ plugins {
 
 android {
     namespace 'de.xera.applight'
-    compileSdk 33
+    compileSdk 34
 
     defaultConfig {
         applicationId "de.xera.applight"
