@@ -10,12 +10,12 @@ import android.bluetooth.le.ScanCallback;
 import android.bluetooth.le.ScanResult;
 import android.content.Context;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.location.LocationManager;
 import android.os.Bundle;
 import android.os.Handler;
 import android.util.Log;
-import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
 
@@ -38,53 +38,47 @@ public class DeviceScanActivity extends AppCompatActivity {
     private BluetoothLeScanner bluetoothLeScanner;
     private LocationManager locationManager;
     private static final long SCAN_PERIOD = 3000;
-
     String[] permissions;
-
     private Button searchDevices;
+    SharedPreferences sharedPref;
+    private boolean scanning = false;
+    private final Handler handler = new Handler();
+
+    private static final String LAST_BLE_DEVICE_ADDRESS = "lastBleDeviceAddress";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_device_scan);
-        permissions = new String[]{Manifest.permission.BLUETOOTH_ADMIN, Manifest.permission.BLUETOOTH_CONNECT, Manifest.permission.BLUETOOTH_SCAN, Manifest.permission.ACCESS_COARSE_LOCATION, Manifest.permission.ACCESS_FINE_LOCATION};
-        Log.d("OnCreate()", "Permissions check:");
-        if (ContextCompat.checkSelfPermission(DeviceScanActivity.this, Manifest.permission.BLUETOOTH_CONNECT) != PackageManager.PERMISSION_GRANTED) {
-            Log.d("OnCreate()", "Permissions missing BLUETOOTH_CONNECT");
-            ActivityCompat.requestPermissions(this, permissions, 1);
-        } else if (ContextCompat.checkSelfPermission(DeviceScanActivity.this, Manifest.permission.BLUETOOTH_SCAN) != PackageManager.PERMISSION_GRANTED) {
-            Log.d("OnCreate()", "Permissions missing BLUETOOTH_SCAN");
-            ActivityCompat.requestPermissions(this, permissions, 1);
-        } else if (ContextCompat.checkSelfPermission(DeviceScanActivity.this, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
-            Log.d("OnCreate()", "Permissions missing ACCESS_COARSE_LOCATION");
-            ActivityCompat.requestPermissions(this, permissions, 1);
-        } else if (ContextCompat.checkSelfPermission(DeviceScanActivity.this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
-            Log.d("OnCreate()", "Permissions missing ACCESS_FINE_LOCATION");
-            ActivityCompat.requestPermissions(this, permissions, 1);
-        } else {
-            Log.d("Permissions", "Permissions granted.");
-        }
+        checkForPermissions();
 
         bluetoothManager = (BluetoothManager) getSystemService(Context.BLUETOOTH_SERVICE);
         bluetoothAdapter = bluetoothManager.getAdapter();
+
+        sharedPref = this.getPreferences(Context.MODE_PRIVATE);
+        String lastDeviceAddress = sharedPref.getString(LAST_BLE_DEVICE_ADDRESS, "none");
+        if(!lastDeviceAddress.equals("none")) {
+            BluetoothDevice lastBluetoothDevice = bluetoothAdapter.getRemoteDevice(lastDeviceAddress);
+
+            Intent intent = new Intent(getApplicationContext(), HomeActivity.class);
+            intent.putExtra("device", lastBluetoothDevice);
+            startActivity(intent);
+        }
+        setContentView(R.layout.activity_device_scan);
         bluetoothLeScanner = bluetoothAdapter.getBluetoothLeScanner();
         locationManager = (LocationManager) getSystemService(Context.LOCATION_SERVICE);
 
         searchDevices = findViewById(R.id.searchDevicesButton);
 
-        searchDevices.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                if (!bluetoothAdapter.isEnabled()) {
-                    Toast.makeText(DeviceScanActivity.this, "Bluetooth ist deaktiviert", Toast.LENGTH_SHORT).show();
-                    return;
-                }
-                if (!locationManager.isLocationEnabled()) {
-                    Toast.makeText(DeviceScanActivity.this, "Standort ist deaktiviert", Toast.LENGTH_SHORT).show();
-                    return;
-                }
-                scanLeDevice();
+        searchDevices.setOnClickListener(view -> {
+            if (!bluetoothAdapter.isEnabled()) {
+                Toast.makeText(DeviceScanActivity.this, "Bluetooth ist deaktiviert", Toast.LENGTH_SHORT).show();
+                return;
             }
+            if (!locationManager.isLocationEnabled()) {
+                Toast.makeText(DeviceScanActivity.this, "Standort ist deaktiviert", Toast.LENGTH_SHORT).show();
+                return;
+            }
+            scanLeDevice();
         });
 
         recyclerView = findViewById(R.id.rv_bluetooth_devices);
@@ -94,25 +88,13 @@ public class DeviceScanActivity extends AppCompatActivity {
         recyclerView.setAdapter(adapter);
     }
 
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        Log.d("Permissions", "Permissions: " + Arrays.toString(grantResults));
-    }
-
-    private boolean scanning = false;
-    private final Handler handler = new Handler();
-
     private void scanLeDevice() {
         if (!scanning) {
             // Stops scanning after a predefined scan period.
-            handler.postDelayed(new Runnable() {
-                @Override
-                public void run() {
-                    scanning = false;
-                    Log.d("DevicesScanActivity", "Scan stopped.");
-                    bluetoothLeScanner.stopScan(leScanCallback);
-                }
+            handler.postDelayed(() -> {
+                scanning = false;
+                Log.d("DevicesScanActivity", "Scan stopped.");
+                bluetoothLeScanner.stopScan(leScanCallback);
             }, SCAN_PERIOD);
 
             scanning = true;
@@ -137,8 +119,37 @@ public class DeviceScanActivity extends AppCompatActivity {
 
 
     public void connectDevice(BluetoothDevice bluetoothDevice) {
+        SharedPreferences.Editor editor = sharedPref.edit();
+        editor.putString(LAST_BLE_DEVICE_ADDRESS, bluetoothDevice.getName());
+        editor.apply();
+
         Intent intent = new Intent(getApplicationContext(), HomeActivity.class);
         intent.putExtra("device", bluetoothDevice);
         startActivity(intent);
+    }
+
+    private void checkForPermissions(){
+        permissions = new String[]{Manifest.permission.BLUETOOTH_ADMIN, Manifest.permission.BLUETOOTH_CONNECT, Manifest.permission.BLUETOOTH_SCAN, Manifest.permission.ACCESS_COARSE_LOCATION, Manifest.permission.ACCESS_FINE_LOCATION};
+        Log.d("OnCreate()", "Permissions check:");
+        if (ContextCompat.checkSelfPermission(DeviceScanActivity.this, Manifest.permission.BLUETOOTH_CONNECT) != PackageManager.PERMISSION_GRANTED) {
+            Log.d("OnCreate()", "Permissions missing BLUETOOTH_CONNECT");
+            ActivityCompat.requestPermissions(this, permissions, 1);
+        } else if (ContextCompat.checkSelfPermission(DeviceScanActivity.this, Manifest.permission.BLUETOOTH_SCAN) != PackageManager.PERMISSION_GRANTED) {
+            Log.d("OnCreate()", "Permissions missing BLUETOOTH_SCAN");
+            ActivityCompat.requestPermissions(this, permissions, 1);
+        } else if (ContextCompat.checkSelfPermission(DeviceScanActivity.this, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+            Log.d("OnCreate()", "Permissions missing ACCESS_COARSE_LOCATION");
+            ActivityCompat.requestPermissions(this, permissions, 1);
+        } else if (ContextCompat.checkSelfPermission(DeviceScanActivity.this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+            Log.d("OnCreate()", "Permissions missing ACCESS_FINE_LOCATION");
+            ActivityCompat.requestPermissions(this, permissions, 1);
+        } else {
+            Log.d("Permissions", "Permissions granted.");
+        }
+    }
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        Log.d("Permissions", "Permissions: " + Arrays.toString(grantResults));
     }
 }
